Move remnants of Dataflow runner from SDK to its module
